test: split channel suites by concern
