+++ tests/unit/channels/test_rss_fetch.py
@@ -64,13 +64,20 @@ def test_rss_identity_uses_entry_id():
 
 @pytest.mark.asyncio
 async def test_fetch_200_returns_items_and_advances_cursor():
-    http = _Http(_Resp(200, text=_RSS, headers={"ETag": 'W/"v2"', "Last-Modified": "Wed, 01 Jul 2026 00:00:00 GMT"}))
+    headers = {
+        "ETag": 'W/"v2"',
+        "Last-Modified": "Wed, 01 Jul 2026 00:00:00 GMT",
+    }
+    http = _Http(_Resp(200, text=_RSS, headers=headers))
     ctx = FetchContext(config={"feed_url": "https://x/feed"}, params={}, cursor=None, http=http)
 
     result = await RSSChannel().fetch(ctx)
 
     assert [i["id"] for i in result.items] == ["id-a", "id-b"]
-    assert result.next_cursor == {"etag": 'W/"v2"', "last_modified": "Wed, 01 Jul 2026 00:00:00 GMT"}
+    assert result.next_cursor == {
+        "etag": 'W/"v2"',
+        "last_modified": "Wed, 01 Jul 2026 00:00:00 GMT",
+    }
     assert result.has_more is False
 
 
@@ -113,57 +120,6 @@ async def test_fetch_304_no_new_items_keeps_cursor_and_sends_conditional():
     assert http.calls[0][1]["headers"]["If-None-Match"] == 'W/"v1"'
 
 
-# ── AUDIT C13: gateway statuses classify retryable, other 4xx stay permanent ──
-
-class _StatusErrorResp:
-    """A response whose raise_for_status() raises a REAL httpx.HTTPStatusError
-    (unlike this file's own _Resp fake, whose raise_for_status raises a bare
-    RuntimeError) — needed to exercise fetch()'s httpx.HTTPStatusError
-    classification branch, which _Resp never triggers."""
-
-    def __init__(self, status_code):
-        import httpx
-
-        self.status_code = status_code
-        self.headers = {}
-        self._exc = httpx.HTTPStatusError(
-            message=f"HTTP {status_code}",
-            request=MagicMock(),
-            response=MagicMock(status_code=status_code, text=f"error {status_code}"),
-        )
-
-    def raise_for_status(self):
-        raise self._exc
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("status", [502, 503, 504, 520, 522, 524])
-async def test_fetch_gateway_status_classified_retryable(status):
-    """fetch()'s response.raise_for_status() used to be bare (no try/except
-    at all) — any status error, gateway or not, propagated as a raw
-    httpx.HTTPStatusError with no retry-classification hint. Now a gateway
-    status must raise ChannelFetchError(error_type="RetryableHTTPStatus")."""
-    http = _Http(_StatusErrorResp(status))
-    ctx = FetchContext(config={"feed_url": "https://x/feed"}, params={}, cursor=None, http=http)
-
-    with pytest.raises(ChannelFetchError) as exc_info:
-        await RSSChannel().fetch(ctx)
-
-    assert exc_info.value.error_type == "RetryableHTTPStatus"
-
-
-@pytest.mark.asyncio
-async def test_fetch_client_404_classified_permanent():
-    """A genuine 4xx (not 408/429) stays permanent."""
-    http = _Http(_StatusErrorResp(404))
-    ctx = FetchContext(config={"feed_url": "https://x/feed"}, params={}, cursor=None, http=http)
-
-    with pytest.raises(ChannelFetchError) as exc_info:
-        await RSSChannel().fetch(ctx)
-
-    assert exc_info.value.error_type == "PermanentHTTPStatus"
-
-
 @pytest.mark.asyncio
 async def test_fetch_bozo_feed_raises_error_type_mapped_to_schema_drift():
     """WIRING_GAP_LEDGER W1: fetch()'s bozo branch must carry error_type (from
@@ -202,7 +158,14 @@ async def test_run_channel_drives_rss_and_persists_cursor():
     http = _Http(_Resp(200, text=_RSS, headers={"ETag": 'W/"v2"'}))
     store = InMemoryCursorStore()
 
-    items = (await run_channel(source, {}, cursor_store=store, channel=RSSChannel(), http=http)).items
+    result = await run_channel(
+        source,
+        {},
+        cursor_store=store,
+        channel=RSSChannel(),
+        http=http,
+    )
+    items = result.items
 
     assert [i["id"] for i in items] == ["id-a", "id-b"]
     # Incremental channel: the runner persisted the advanced cursor.
